refactor(example): remove unused decoder, rename packages
