+++ example/review/src/main/scala/review/State.scala
@@ -1,16 +1,12 @@
-package core
+package review
 
-import io.circe.generic.semiauto.deriveDecoder
 import sangria.schema._
-import io.circe.Decoder
 
 case class State(id: Int)
 
 object State {
 
   import sangria.federation.Directives._
-
-  implicit val decoder: Decoder[State] = deriveDecoder
 
   implicit val stateSchema =
     ObjectType(
